requirements.txt
selenium
beautifulsoup4
pandas
tqdm
aiohttp
pyarrow
python-dotenv 
grpcio
protobuf
sqlalchemy
psycopg2
